backend/pyproject.toml
[project]
name = "forge-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "chromadb>=1.5.5",
    "cryptography>=46.0.5",
    "fastapi[standard]>=0.135.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "openai>=2.29.0",
    "pwdlib[bcrypt]>=0.3.0",
    "pydantic-settings>=2.13.1",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "httpx>=0.28.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.7",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
